Fix#1268 rename folder to kebab-case
