+++ packages/docs/src/components/DocsApi/ApiDocs.vue
@@ -161,7 +161,7 @@ import { DefineComponent } from 'vue'
 import { ManualApiOptions } from './ManualApiOptions'
 import ApiDocsPropsRow from './ApiDocsPropsRow.vue'
 import { getApiTableData } from './api-docs-helpers'
-import MarkdownView from '../MarkdownView/MarkdownView.vue'
+import MarkdownView from '../markdown-view/MarkdownView.vue'
 import { defaultApiOptions } from './default-api-options'
 import DocsTable from '../DocsTable/DocsTable.vue'
 import { merge } from 'lodash'
