+++ packages/docs/src/pages/MarkdownTest.vue
@@ -5,7 +5,7 @@
 <script lang="ts">
 // @ts-ignore
 import { Options, Vue } from 'vue-class-component'
-import MdView from '../components/MarkdownView/MarkdownView.vue'
+import MdView from '../components/markdown-view/MarkdownView.vue'
 
 @Options({
   name: 'MarkdownTest',
